requirements.txt
Flask==2.3.3
gunicorn==21.2.0
openpyxl==3.1.2
pandas==1.5.3
numpy==1.24.3
playwright==1.39.0
xlrd==2.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
openpyxl-stubs==0.1.25
